+++ lib/user_agent/browsers.rb
@@ -14,6 +14,7 @@ require 'user_agent/browsers/apple_podcasts'
 require 'user_agent/browsers/apple_watch/base'
 require 'user_agent/browsers/apple_watch/atc'
 require 'user_agent/browsers/apple_watch/pod_2_watch'
+require 'user_agent/browsers/audible'
 require 'user_agent/browsers/black_berry'
 require 'user_agent/browsers/castbox'
 require 'user_agent/browsers/chrome'
@@ -80,6 +81,7 @@ class UserAgent
       Chrome,
       ITunes,
       PlayStation,
+      Audible,
       BlackBerry,
       Webkit,
       Gecko,