+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 1.2.0
-* Add browsers support for Wondery, BlackBerry, general Android
+* Add browsers support for Audible, BlackBerry, Wondery, general Android
 * Improve OkHttp
 
 # 1.1.0